Made database.py package from the single module and repaired hdf5.py. Works with cmf_lumped
